+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - OPI-104: Fixed card badge's positioning
 
+### Changed
+
+- OPI-105: Changed "Affects user profile" translation text
+
 ## [0.3.3] - 2022-08-15
 
 ### Fixed